+++ src/__tests__/views/TransactionsPage/TransactionsDetail.test.js
@@ -4,7 +4,7 @@ import {GenericDetail} from "../../../views/GenericPage/GenericDetail";
 
 describe("Test Transactions Detail", () => {
   it("Test can show single detail", async () => {
-    renderWithProviders(<GenericDetail id={1} />);
+    renderWithProviders(<GenericDetail id={99} />);
     await waitFor(() => {
       const purpose = screen.getByText(/december 2022 - salary/i);
       expect(purpose).toBeInTheDocument();