+++ src/__tests__/views/TransactionsPage/TransactionsPage.test.js
@@ -15,25 +15,6 @@ describe("Test TransactionsPage", () => {
     });
   });
 
-  it("Test can manage zero transactions", async () => {
-
-    const zeroItems = {
-      content: []
-    };
-
-    server.use(
-      rest.get(`${process.env.REACT_APP_BACKEND_URL}/bank-account-transactions`, (req, res, ctx) => {
-        return res(ctx.json(zeroItems));
-      })
-    );
-
-    renderWithProviders(<GenericPage entity={ENTITY_TRANSACTIONS} />);
-    await waitFor(() => {
-      const noData = screen.getByText(/There are no records to display/i);
-      expect(noData).toBeInTheDocument();
-    });
-  });
-
   it("Test can manage change page", async () => {
 
     renderWithProviders(<GenericPage entity={ENTITY_TRANSACTIONS} />);