+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Upgrade dependencies
 + Refactor tests
 + Upgrade dependencies
++ Add QRCode to enable MFA
 
 ## 1.0.0
 + Upgrade dependencies + reorganize Docker