+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 1.1.0
 + Upgrade dependencies
 + Refactor tests
 + Upgrade dependencies